31/10/2022

+++ spring-boot-server/src/main/java/com/CMSBACK/models/Role.java
@@ -10,7 +10,7 @@ import org.hibernate.envers.Audited;
 public class Role {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@Enumerated(EnumType.STRING)
 	@Column(length = 20)
@@ -24,11 +24,11 @@ public class Role {
 		this.name = name;
 	}
 
-	public Integer getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(long id) {
 		this.id = id;
 	}
